requirements.txt
# dash is required to call `build:py`
dash[dev]>=2.18.1

setuptools~=75.1.0
python-dateutil~=2.9.0.post0